kipps-ai-analysis/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
textblob==0.17.1
vaderSentiment==3.3.2
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
gunicorn==21.2.0